requirements.txt
absl-py==2.3.1
antlr4-python3-runtime==4.9.3
attrs==25.4.0
black==25.11.0
cattrs==25.3.0
click==8.3.0
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
execnet==2.1.2
expecttest==0.3.0
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.10.0
grpcio==1.76.0
humanfriendly==10.0
hydra-core==1.3.2
hypothesis==6.147.0
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.2
kgb==7.2
kiwisolver==1.4.9
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.5
numpy==2.3.4
omegaconf==2.3.0
onnx==1.19.1
onnxruntime==1.23.2
packaging==25.0
pandas==2.3.3
parameterized==0.9.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
protobuf==6.33.1
pyaml==25.7.0
pycocotools==2.0.10
Pygments==2.19.2
pyparsing==3.2.5
pytest==9.0.1
pytest-json-report==1.5.0
pytest-metadata==3.1.1
pytest-rerunfailures==15.1
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
pytokens==0.3.0
pytz==2025.2
pyvww==0.1.1
PyYAML==6.0.3
rich==14.2.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
scikit-learn==1.7.1
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
sortedcontainers==2.4.0
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.1.3
coremltools==9.0

torchao==0.13.0

# Windows (CUDA)
# install with "pip install -r requirements.txt --extra-index-url https://download.pytorch.org/whl/cu129"
torch==2.8.0+cu129; platform_system == "Windows"
torchaudio==2.8.0+cu129; platform_system == "Windows"
torchvision==0.23.0+cu129; platform_system == "Windows"

# macOS (CPU/MPS)
# install with "pip install -r requirements.txt"
torch==2.9.0; platform_system == "Darwin"
torchaudio==2.9.0; platform_system == "Darwin"
torchvision==0.24.0; platform_system == "Darwin"

